\name{CoSeg-package}
\alias{CoSeg-package}
\alias{CoSeg}
\docType{package}
\title{
\packageTitle{CoSeg}
}
\description{
\packageDescription{CoSeg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CoSeg}
\packageIndices{CoSeg}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{CoSeg}

Maintainer: \packageMaintainer{CoSeg}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{

}
\examples{
#this goes in seealso
#%~~ Optional links to other man pages, e.g. ~~
#%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~

#~~ simple examples of the most important functions ~~
}
